refactored main method to shift accepting input to it's own service, closes #7
